refactor: extract demo rush and half volley stats accumulators
