+++ src/stats/calculators/mod.rs
@@ -77,6 +77,8 @@ pub mod center_stats;
 pub use center_stats::*;
 pub mod demo;
 pub use demo::*;
+pub mod demo_stats;
+pub use demo_stats::*;
 mod flip_reset;
 pub use flip_reset::*;
 mod flip_reset_tuning_set;
@@ -105,6 +107,8 @@ pub mod half_flip_stats;
 pub use half_flip_stats::*;
 pub mod half_volley;
 pub use half_volley::*;
+pub mod half_volley_stats;
+pub use half_volley_stats::*;
 pub mod match_stats;
 pub use match_stats::*;
 pub mod movement;
@@ -137,6 +141,8 @@ pub mod rotation;
 pub use rotation::*;
 pub mod rush;
 pub use rush::*;
+pub mod rush_stats;
+pub use rush_stats::*;
 pub mod settings;
 pub use settings::*;
 pub mod speed_flip;
